MovieHubNovember/requirement.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
geographiclib==2.0
geopy==2.4.1
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.2
